restruct demo
